+++ addons/web/static/src/core/debug/debug_context.js
@@ -1,4 +1,5 @@
 /** @odoo-module **/
+
 import { registry } from "../registry";
 import { memoize } from "../utils/functions";
 import { useEffect } from "@web/core/utils/hooks";
@@ -78,9 +79,12 @@ export function useEnvDebugContext() {
 }
 
 export function useDebugCategory(category, context = {}) {
-    const debugContext = useEnvDebugContext();
-    useEffect(
-        () => debugContext.activateCategory(category, context),
-        () => []
-    );
+    const env = useEnv();
+    if (env.debug) {
+        const debugContext = useEnvDebugContext();
+        useEffect(
+            () => debugContext.activateCategory(category, context),
+            () => []
+        );
+    }
 }
