+++ addons/web/static/src/legacy/debug_manager.js
@@ -1,212 +1,15 @@
 /** @odoo-module **/
 
-import { Dialog } from "../core/dialog/dialog";
-import { formatDateTime, parseDateTime } from "../core/l10n/dates";
-import { _lt } from "../core/l10n/translation";
-import { registry } from "../core/registry";
+import { _lt } from "@web/core/l10n/translation";
+import { Dialog } from "@web/core/dialog/dialog";
+import { formatDateTime, parseDateTime } from "@web/core/l10n/dates";
 import { formatMany2one } from "@web/fields/formatters";
+import { registry } from "@web/core/registry";
 
-const { hooks, tags } = owl;
+const { hooks } = owl;
 const { useState } = hooks;
 
-function editModelDebug(env, title, model, id) {
-    return env.services.action.doAction({
-        res_model: model,
-        res_id: id,
-        name: title,
-        type: "ir.actions.act_window",
-        views: [[false, "form"]],
-        view_mode: "form",
-        target: "new",
-        flags: { action_buttons: true, headless: true },
-    });
-}
-
-// Action items
-
-function actionSeparator() {
-    return {
-        type: "separator",
-        sequence: 100,
-    };
-}
-
-function accessSeparator({ accessRights, action }) {
-    const { canSeeModelAccess, canSeeRecordRules } = accessRights;
-    if (!action.res_model || (!canSeeModelAccess && !canSeeRecordRules)) {
-        return null;
-    }
-    return {
-        type: "separator",
-        sequence: 200,
-    };
-}
-
-function editAction({ action, env }) {
-    if (!action.id) {
-        return null;
-    }
-    const description = env._t("Edit Action");
-    return {
-        type: "item",
-        description,
-        callback: () => {
-            editModelDebug(env, description, action.type, action.id);
-        },
-        sequence: 110,
-    };
-}
-
-function viewFields({ action, env }) {
-    if (!action.res_model) {
-        return null;
-    }
-    const description = env._t("View Fields");
-    return {
-        type: "item",
-        description,
-        callback: async () => {
-            const modelId = (
-                await env.services.orm.search("ir.model", [["model", "=", action.res_model]], {
-                    limit: 1,
-                })
-            )[0];
-            env.services.action.doAction({
-                res_model: "ir.model.fields",
-                name: description,
-                views: [
-                    [false, "list"],
-                    [false, "form"],
-                ],
-                domain: [["model_id", "=", modelId]],
-                type: "ir.actions.act_window",
-                context: {
-                    default_model_id: modelId,
-                },
-            });
-        },
-        sequence: 120,
-    };
-}
-
-function manageFilters({ action, env }) {
-    if (!action.res_model) {
-        return null;
-    }
-    const description = env._t("Manage Filters");
-    return {
-        type: "item",
-        description,
-        callback: () => {
-            // manage_filters
-            env.services.action.doAction({
-                res_model: "ir.filters",
-                name: description,
-                views: [
-                    [false, "list"],
-                    [false, "form"],
-                ],
-                type: "ir.actions.act_window",
-                context: {
-                    search_default_my_filters: true,
-                    search_default_model_id: action.res_model,
-                },
-            });
-        },
-        sequence: 130,
-    };
-}
-
-function technicalTranslation({ action, env }) {
-    if (!action.res_model) {
-        return null;
-    }
-    return {
-        type: "item",
-        description: env._t("Technical Translation"),
-        callback: async () => {
-            const result = await env.services.orm.call(
-                "ir.translation",
-                "get_technical_translations",
-                [action.res_model]
-            );
-            env.services.action.doAction(result);
-        },
-        sequence: 140,
-    };
-}
-
-function viewAccessRights({ accessRights, action, env }) {
-    if (!action.res_model || !accessRights.canSeeModelAccess) {
-        return null;
-    }
-    const description = env._t("View Access Rights");
-    return {
-        type: "item",
-        description,
-        callback: async () => {
-            const modelId = (
-                await env.services.orm.search("ir.model", [["model", "=", action.res_model]], {
-                    limit: 1,
-                })
-            )[0];
-            env.services.action.doAction({
-                res_model: "ir.model.access",
-                name: description,
-                views: [
-                    [false, "list"],
-                    [false, "form"],
-                ],
-                domain: [["model_id", "=", modelId]],
-                type: "ir.actions.act_window",
-                context: {
-                    default_model_id: modelId,
-                },
-            });
-        },
-        sequence: 210,
-    };
-}
-
-function viewRecordRules({ accessRights, action, env }) {
-    if (!action.res_model || !accessRights.canSeeRecordRules) {
-        return null;
-    }
-    const description = env._t("Model Record Rules");
-    return {
-        type: "item",
-        description: env._t("View Record Rules"),
-        callback: async () => {
-            const modelId = (
-                await env.services.orm.search("ir.model", [["model", "=", action.res_model]], {
-                    limit: 1,
-                })
-            )[0];
-            env.services.action.doAction({
-                res_model: "ir.rule",
-                name: description,
-                views: [
-                    [false, "list"],
-                    [false, "form"],
-                ],
-                domain: [["model_id", "=", modelId]],
-                type: "ir.actions.act_window",
-                context: {
-                    default_model_id: modelId,
-                },
-            });
-        },
-        sequence: 220,
-    };
-}
-
-class FieldViewGetDialog extends Dialog {}
-FieldViewGetDialog.props = Object.assign({}, Dialog.props, {
-    arch: { type: String },
-    close: Function,
-});
-FieldViewGetDialog.bodyTemplate = tags.xml`<pre t-esc="props.arch"/>`;
-FieldViewGetDialog.title = _lt("Fields View Get");
+const debugRegistry = registry.category("debug");
 
 class GetMetadataDialog extends Dialog {
     setup() {
@@ -243,7 +46,6 @@ class GetMetadataDialog extends Dialog {
 }
 GetMetadataDialog.bodyTemplate = "web.DebugMenu.getMetadataBody";
 GetMetadataDialog.title = _lt("View Metadata");
-
 class SetDefaultDialog extends Dialog {
     setup() {
         super.setup();
@@ -389,66 +191,7 @@ SetDefaultDialog.bodyTemplate = "web.DebugMenu.setDefaultBody";
 SetDefaultDialog.footerTemplate = "web.DebugMenu.SetDefaultFooter";
 SetDefaultDialog.title = _lt("Set Default");
 
-function viewSeparator() {
-    return {
-        type: "separator",
-        sequence: 300,
-    };
-}
-
-function fieldsViewGet({ component, env }) {
-    return {
-        type: "item",
-        description: env._t("Fields View Get"),
-        callback: () => {
-            const props = {
-                arch: component.props.viewInfo.arch,
-            };
-            env.services.dialog.add(FieldViewGetDialog, props);
-        },
-        sequence: 340,
-    };
-}
-
-export function editView({ accessRights, action, component, env }) {
-    if (!accessRights.canEditView) {
-        return null;
-    }
-    let { view_id: viewId, type: viewType } = component.props.viewInfo;
-    viewType = viewType === "tree" ? "list" : viewType;
-    const displayName = action.views.find((v) => v.type === viewType).name.toString();
-    const description = env._t("Edit View: ") + displayName;
-    return {
-        type: "item",
-        description,
-        callback: () => {
-            editModelDebug(env, description, "ir.ui.view", viewId);
-        },
-        sequence: 350,
-    };
-}
-
-export function editSearchView({ accessRights, component, env }) {
-    if (!accessRights.canEditView) {
-        return null;
-    }
-    const description = env._t("Edit SearchView");
-    return {
-        type: "item",
-        description,
-        callback: () => {
-            editModelDebug(
-                env,
-                description,
-                "ir.ui.view",
-                component.props.viewParams.controlPanelFieldsView.view_id
-            );
-        },
-        sequence: 360,
-    };
-}
-
-// Form view itemss
+// Form view items
 
 function setDefaults({ action, component, env }) {
     return {
@@ -514,26 +257,6 @@ function manageAttachments({ action, component, env }) {
         sequence: 330,
     };
 }
-
-const debugRegistry = registry.category("debug");
-
-debugRegistry
-    .category("action")
-    .add("actionSeparator", actionSeparator)
-    .add("editAction", editAction)
-    .add("viewFields", viewFields)
-    .add("manageFilters", manageFilters)
-    .add("technicalTranslation", technicalTranslation)
-    .add("accessSeparator", accessSeparator)
-    .add("viewAccessRights", viewAccessRights)
-    .add("viewRecordRules", viewRecordRules);
-
-debugRegistry
-    .category("view")
-    .add("viewSeparator", viewSeparator)
-    .add("fieldsViewGet", fieldsViewGet)
-    .add("editView", editView)
-    .add("editSearchView", editSearchView);
 
 debugRegistry
     .category("form")
